新增工具类

+++ AndroidNetOptionsFromLorenWang/src/main/java/com/lorenwang/netoptions/android/NetworkOptionsCallback.java
@@ -17,7 +17,7 @@ public abstract class NetworkOptionsCallback {
     public abstract void fail(int code, Object object);
     public abstract void error(int code, Object object);
 
-    public void progress(double progress) {
+    public void progress(int progress) {
 
     }
 }
